pyproject.toml
[project]
name = "fast-api-template"
version = "0.1.0"
description = "FastAPI template with async SQLModel, PostgreSQL, Alembic, JWT auth, and Docker"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "alembic~=1.16.5",
    "asyncpg~=0.30.0",
    "bcrypt~=4.0.0",
    "email-validator~=2.0.0",
    "fastapi[standard]~=0.118.0",
    "greenlet~=3.2.4",
    "pydantic-settings~=2.11.0",
    "PyJWT~=2.10.0",
    "slowapi~=0.1.9",
    "python-dotenv~=1.1.1",
    "sqlmodel[asyncio]~=0.0.25",
    "uvicorn[standard]~=0.37.0",
]

[dependency-groups]
dev = [
    "pytest~=8.0.0",
    "pytest-asyncio~=0.23.0",
    "httpx~=0.27.0",
    "faker~=24.0.0",
    "pre-commit~=3.8.0",
    "ruff~=0.6.0",
    "mypy~=1.11.0",
]

[tool.uv]
default-groups = ["dev"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
# Memory-efficient option: --disable-warnings reduces output and memory usage
addopts = "-v --tb=short --disable-warnings"

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    ".mypy_cache",
    ".ruff_cache",
    "migrations",
    "*.pyc",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["domains", "api", "db", "middleware", "utils"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.13"
mypy_path = "src"
explicit_package_bases = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "alembic.*",
    "tests.*",
]
ignore_missing_imports = true
